+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Development
 ***********
 
 - Make parameters TEMPERATURE_AIR_MAX_200 and TEMPERATURE_AIR_MIN_200 summarizable/interpolatable
+- Add streamlit app for DWD climate stations
 
 0.60.0 (16.09.2023)
 *******************